moved stats inside eda
